requirements.txt
djangorestframework== 3.14.0